requirements.txt
streamlit
scikit-learn
numpy
pillow
tensorflow  
matplotlib
